Move nask executable into libnask dir
