+++ 20GO/libnask/nasktest.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <memory>
 #include <sstream>
-#include "nask.hpp"
+#include "libnask.hpp"
 
 template <class T> void address_cout(T t);
 template <class T> std::string address_get(T t);
